requirements.txt
flask
discord.py
praw

